UI for deleted and sactioned products
