+++ pages/super-admin/deleted-products/deletedProducts.ts
@@ -1,10 +1,10 @@
-import { DeletedProducts } from '../../@types';
+import { DeletedProducts } from '../../../@types';
 
 const products: DeletedProducts[] = [
   {
     name: 'Learning Design 101',
     vendor: 'Okereke James',
-    id: '123456',
+    id: 123456,
     dateAdded: new Date('2021-09-01'),
     dateDeleted: new Date('2021-09-10'),
     status: 'deleted',
@@ -12,7 +12,7 @@ const products: DeletedProducts[] = [
   {
     name: 'Learning Design 101',
     vendor: 'Chukwu Chinaza',
-    id: '123456',
+    id: 123456,
     dateAdded: new Date('2021-09-02'),
     dateDeleted: new Date('2021-09-11'),
     status: 'deleted',
@@ -20,7 +20,7 @@ const products: DeletedProducts[] = [
   {
     name: 'Learning Design 101',
     vendor: 'Mark Essien',
-    id: '123456',
+    id: 123456,
     dateAdded: new Date('2021-09-03'),
     dateDeleted: new Date('2021-09-12'),
     status: 'deleted',
@@ -28,7 +28,7 @@ const products: DeletedProducts[] = [
   {
     name: 'Learning Design 101',
     vendor: 'Shobande Abraham',
-    id: '123456',
+    id: 123456,
     dateAdded: new Date('2021-09-01'),
     dateDeleted: new Date('2021-09-10'),
     status: 'deleted',
@@ -36,7 +36,7 @@ const products: DeletedProducts[] = [
   {
     name: 'Learning Design 101',
     vendor: 'John Paul',
-    id: '123456',
+    id: 123456,
     dateAdded: new Date('2021-09-02'),
     dateDeleted: new Date('2021-09-11'),
     status: 'deleted',
@@ -44,7 +44,7 @@ const products: DeletedProducts[] = [
   {
     name: 'Learning Design 101',
     vendor: 'Bakare Femi',
-    id: '123456',
+    id: 123456,
     dateAdded: new Date('2021-09-03'),
     dateDeleted: new Date('2021-09-12'),
     status: 'deleted',
@@ -52,7 +52,7 @@ const products: DeletedProducts[] = [
   {
     name: 'Learning Design 101',
     vendor: 'Gustavo Silas',
-    id: '123456',
+    id: 123456,
     dateAdded: new Date('2021-09-01'),
     dateDeleted: new Date('2021-09-10'),
     status: 'deleted',
