+++ pages/super-admin/deleted-products/index.tsx
@@ -1,11 +1,11 @@
+import { useState } from 'react';
 import { More } from 'iconsax-react';
-import products from './sanctionedProducts';
-import Pagination from '../view-components/super-admin/pagination';
+import products from './deletedProducts';
 import SuperAdminNavbar from '@modules/super-admin/components/navigations/SuperAdminNavbar';
 import SearchProduct from '@modules/super-admin/components/product-listing/searchProduct';
-import { useState } from 'react';
+import Pagination from '../../view-components/super-admin/pagination';
 
-const SanctionedProducts = () => {
+const DeletedProducts = () => {
   const [searchVal, setSearchVal] = useState('');
   const [filteredProduct, setFilteredProducts] = useState(products);
 
@@ -25,15 +25,16 @@ const SanctionedProducts = () => {
   return (
     <>
       <SuperAdminNavbar />
-
-      <div className="m-6 font-manropeL border border-gray-100">
+      <div className="m-6 font-manropeL border border-custom-color1">
+        {/* <ProductListingTable /> */}
         {/* Heading */}
         <div className="py-3 px-4 flex flex-col md:flex-row justify-between md:items-center gap-4">
           <div>
-            <h2 className="text-lg font-medium text-custom-color10">Sanctioned Products</h2>
-            <p className="text-custom-color2 text-sm">List of all sanctioned products and their details</p>
+            <h2 className="text-lg font-medium text-custom-color10">Deleted Products</h2>
+            <p className="text-custom-color2 text-sm">List of all deleted products and their details</p>
           </div>
           <div>
+            {/* input for search */}
             <SearchProduct handleSearchChange={handleSubmit} />
           </div>
         </div>
@@ -42,24 +43,24 @@ const SanctionedProducts = () => {
           <thead>
             {/* Table Headers */}
             <tr>
-              <th className="text-custom-color2 text-xs font-normal leading-[18px] px-6 py-3 gap-3 text-left flex  items-center">
+              <th className="text-custom-color2 text-xs font-normal leading-5 px-6 py-3 gap-3 text-left flex  items-center">
                 <input type="checkbox" />
-                <samp>Product Name</samp>
+                <span>Product Name </span>
               </th>
-              <th className="text-custom-color2 text-xs font-normal leading-[18px] px-3 py-3 gap-3">Vendor</th>
-              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-[18px] px-3 py-3 gap-3 ">
+              <th className="text-custom-color2 text-xs font-normal leading-5 px-3 py-3 gap-3">Vendor</th>
+              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-5 px-3 py-3 gap-3 ">
                 ID
               </th>
-              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-[18px] px-3 py-3 gap-3">
+              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-5 px-3 py-3 gap-3">
                 Date Added
               </th>
-              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-[18px] px-3 py-3 gap-3">
+              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-5 px-3 py-3 gap-3">
                 Date Deleted
               </th>
-              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-[18px] px-3 py-3 gap-3">
+              <th className="hidden md:table-cell text-custom-color2 text-xs font-normal leading-5 px-3 py-3 gap-3">
                 Status
               </th>
-              <th className="hidden lg:table-cell text-custom-color2 text-xs font-normal leading-[18px] px-3 py-3 gap-3">
+              <th className="hidden lg:table-cell text-custom-color2 text-xs font-normal leading-5 px-3 py-3 gap-3">
                 Action
               </th>
             </tr>
@@ -67,8 +68,8 @@ const SanctionedProducts = () => {
           <tbody>
             {/* Listed Products */}
             {products.map((product, index) => (
-              <tr className="border-t border-gray-200" key={index}>
-                <td className="text-xs tracking-wider lg:tracking-wide font-manropeL lg:text-base text-custom-color2 px-6 py-4 items-center gap-6 self-stretch flex">
+              <tr className="border-t border-custom-color1" key={index}>
+                <td className="text-xs tracking-wider lg:tracking-wide font-manropeL lg:text-base text-custom-color10 px-6 py-4 items-center gap-6 self-stretch flex">
                   <input type="checkbox" />
 
                   {product.name}
@@ -90,9 +91,9 @@ const SanctionedProducts = () => {
                   {(product.dateDeleted.getFullYear() % 100).toString().padStart(2, '0')}
                 </td>
                 <td className="hidden md:table-cell text-xs tracking-wider lg:tracking-wide font-manropeL lg:text-base text-custom-color2 px-6 py-4 text-center">
-                  <div className="hidden mx-auto bg-custom-color40 text-yellow-600 rounded-2xl py-0.5 pl-1.5 pr-2 text-center font-manropeL text-xs font-medium md:flex items-center justify-center gap-2 w-max">
-                    <span className="inline-block w-2 h-2 bg-yellow-600 rounded-full"></span>
-                    <span className="capitalize">{product.status}</span>
+                  <div className="hidden mx-auto bg-pink-120 text-custom-color34 rounded-2xl py-0.5 pl-1.5 pr-2 text-center font-manropeL text-xs font-medium md:flex items-center justify-center gap-2 w-max">
+                    <span className="inline-block w-2 h-2 bg-custom-color34 rounded-full"></span>
+                    <span className=" capitalize">{product.status}</span>
                   </div>
                 </td>
 
@@ -109,4 +110,4 @@ const SanctionedProducts = () => {
   );
 };
 
-export default SanctionedProducts;
+export default DeletedProducts;
